pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "arg_init"
authors = [
    {name = "srfoster65", email = "135555068+srfoster65@users.noreply.github.com"},
]
description = "Resolve argument values from arg, env or default values"
keywords = ["argparse", "init", "environment"]
readme = "readme.md"
requires-python =">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "python-box>=7.0.1",
    "pyyaml>=6.0.1",
]
license = {text = "MIT"}
dynamic = ["version"]

[project.urls]
Documentation = "https://srfoster65.github.io/arg_init/"
Source = "https://github.com/srfoster65/arg_init"


[tool.pdm.version]
source = "file"
path = "src/arg_init/_version.py"

[tool.pdm.dev-dependencies]
lint = [
    "black>=23.9.1",
]
test = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "pyfakefs>=5.3.0",
    "mypy>=1.6.1",
    "types-PyYAML>=6.0.12.12",
    "pydantic>=2.4.2",
    "ruff>=0.1.5",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.3",
    "pymdown-extensions>=10.3",
]


[tool.pylint.basic]
good-names = [
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint."massages-control"]
disable = []


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=src --cov-report term-missing"
testpaths = [
    "tests",
]


[tool.mypy]
plugins = [
  "pydantic.mypy"
]
files = "src/arg_init/"

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

show-error-content = true


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    # "tests/**"
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.8
target-version = "py311"

[tool.ruff.lint]
exclude = [
    "tests/**"
]
# select = [
#     "A",  # prevent using keywords that clobber python builtins
#     "B",  # bugbear: security warnings
#     "E",  # pycodestyle
#     "F",  # pyflakes
#     "ISC",  # implicit string concatenation
#     "UP",  # alert you when better syntax is available in your python version
#     "RUF",  # the ruff developer's own rules
# ]
select = ["ALL"]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# select = ["E4", "E7", "E9", "F"]
ignore = ["ANN002", "ANN003", "ANN101", "ANN102", "D203", "D212", "COM812", "ISC001", "D205"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104"]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


